Create A script to ask a person if they like cats.
This is a program prompt user just confirming if they like cats by entering yes or no. If they enter yes there's some message that will get displayed, and if they enter no there's some other message that will get displayed and if they enter something which is not equal to yes or no then something else will get displayed to the user.