Create A script if the number is divisible by 5 and also even
This script show if the number is divisible by 5 and also even just by calling a method is_even? in the new method using logical operator to check if the number is divisible by 5 and also even. 
There are some test cases to check if a number is divisible by 5 and also even at the bottom after the second method.