Create A program that prompt the user for their age in years and convert the value to an integer to display some message.
This is a script where a user is asked to enter his/her age and converting that user entered value in a integer value. If a user enter his age which is less than 21 there's a message telling something will be displayed and else something other message will be displayed.