Create A program to remove duplicate items from an array
This program shows how to remove duplicate items from an array using two versions, one version is direct with uniq! method and other is not using the uniq! method where the version will loop through the input array and build an output array by pushing items on it depending whether the item is included in the new array or not.