Create A program to play the Foobar game and display n elements of the Foobar pattern!
This problem has its origins in a children's game for learning about division, remainders and fractions. Teachers have schoolchildren sit in a circle, and the person who starts the game says the number "1." Each student in the circle has a turn to say the next number in a pattern. The catch is: for all numbers that are divisible by three, the player should say "foo." For any number divisible by five, the player should say "bar." Numbers divisible by both become "Foobar."